cmd/image/tree: refactor
